% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncbi_taxonomy.R
\name{get_taxid}
\alias{get_taxid}
\title{Retrieve NCBI taxonomic identifiers (taxid) for accessions}
\usage{
get_taxid(sql_path, accessions, show_elapsed = T)
}
\arguments{
\item{sql_path}{Character scalar. Path to the SQLite database generated by
\code{make_database()}, containing the \code{accession2taxid} table.}

\item{accessions}{Character vector of NCBI accessions to be queried.
All accessions should be either versioned or unversioned.}

\item{show_elapsed}{Logical. Whether to display elapsed time information
using \code{show_time()}. Default is \code{TRUE}.}
}
\value{
A \code{data.frame} with two columns:
\describe{
\item{accession}{Input accession identifiers, in the same order as provided.}
\item{taxid}{NCBI taxonomic identifier corresponding to each accession.}
}
}
\description{
Query an SQLite accession-to-taxid database to obtain taxonomic identifiers
corresponding to a vector of NCBI accessions. Both versioned accessions
(e.g. \code{AB123456.1}) and unversioned accessions (e.g. \code{AB123456})
are supported and are detected automatically.
}
\details{
The function returns results in the same order as the input accessions.
Accessions not found in the database will have \code{NA} as their taxid.

The function checks whether the database contains a table named
\code{accession2taxid}. If not found, execution is stopped with an error.
SQL queries are constructed internally using \code{IN (...)} clauses;
therefore, extremely large accession vectors may be limited by SQLite
query length constraints.
}
\examples{
\dontrun{
db <- "accession2taxid.sqlite"
acc <- c("AB123456.1", "NM_001200025.1")
res <- get_taxid(db, acc)
head(res)
}

}
\seealso{
\code{\link{make_database}}
}
